pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infinity-dual-hybrid"
version = "2.1.3"
description = "Production-grade RL framework with Mamba2, Dual-Tier Miras, FAISS LTM, and unified logging"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Dawson Block"}
]
keywords = ["reinforcement-learning", "mamba", "memory", "ppo"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "pyyaml>=6.0",
    "tensorboard>=2.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black",
    "flake8",
    "mypy",
    "ruff>=0.5.0",
    "types-PyYAML>=6.0.0.0",
]
faiss = [
    "faiss-cpu>=1.7.4",
]
mamba = [
    "mamba-ssm>=1.0.0",
]

env = [
    "gymnasium[classic-control]>=0.29.0",
    "transformers>=4.40.0",
]
tensorboard = [
    "tensorboard>=2.10.0",
]
transformers = [
    "transformers>=4.40.0",
]
all = [
    "faiss-cpu>=1.7.4",
    "mamba-ssm>=1.0.0",
    "tensorboard>=2.10.0",
    "gymnasium[classic-control]>=0.29.0",
]

[project.scripts]
infinity-best = "infinity_dual_hybrid.best_checkpoint:main"
infinity-train = "infinity_dual_hybrid.train_cli:main"
infinity-eval = "infinity_dual_hybrid.eval:main"
infinity-eval-runs = "infinity_dual_hybrid.auto_eval_runs:main"
infinity-sweep = "infinity_dual_hybrid.sweep:main"
infinity-report = "infinity_dual_hybrid.report:main"
infinity-smoke = "infinity_dual_hybrid.smoke:smoke_main"

[project.urls]
Homepage = "https://github.com/dawsonblock/INFINITY_DUAL_HYBRID"
Repository = "https://github.com/dawsonblock/INFINITY_DUAL_HYBRID"

[tool.setuptools.packages.find]
where = ["src"]
include = ["infinity_dual_hybrid*", "humanlike_core*"]
exclude = ["tests*", "legacy*", "vendor*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests", "src/humanlike_core/tests"]
python_files = ["test_*.py"]
addopts = "-ra -p no:ddtrace"

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203", "E266", "E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
